1.code/langraph/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-aws>=0.2.0
python-dotenv>=1.0.0
boto3>=1.35.0
pydantic>=2.0.0
